% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destc.R
\name{calculateMetrics}
\alias{calculateMetrics}
\title{calculateMetrics}
\usage{
calculateMetrics(results_df)
}
\arguments{
\item{results_df}{A data frame whose first column contains "Target" values
and other columns are models predictions.}
}
\value{
A data frame with RMSE, MAE and Theil metrics
}
\description{
Calculates a set of metrics (RMSE, MAE, and Theil)
}
\examples{
test_df = data.frame(matrix(ncol=3, nrow=5))
colnames(test_df) = c("Target", "Model_1", "Model_2")
test_df$Target = c(2.7, 2.5, 3.1, 5.6, 5.7)
test_df$Model_1 = c(1.7, 3.5, 2.1, 2.6, 5.1)
test_df$Model_2 = c(3.7, 5.1, 2.4, 2.2, 4.2)
calculateMetrics(test_df)
}
