% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destc.R
\name{createPlot}
\alias{createPlot}
\title{createPlot}
\usage{
createPlot(df, legend_place = "topright", leg_names = c("Target", "Pool"))
}
\arguments{
\item{df}{A data frame whose first column contains "Target" values
and other columns are models predictions.}

\item{legend_place}{Character (default = 'topright'). Legend place.}

\item{leg_names}{Character. Legend names.}
}
\value{
A plot
}
\description{
Generates a graphic with Target and the pool models
}
\examples{
test_df = data.frame(matrix(ncol=20, nrow=100))
colnames(test_df) = c("Target")
test_df$Target = rexp(100, rate = 0.2)
test_df$model_1 = rexp(100, rate = 0.1)
for(model in 2:20){
  test_df[,model] = rexp(100)
 }
  createPlot(test_df)
}
