% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destc.R
\name{ensemblingModels}
\alias{ensemblingModels}
\title{ensemblingModels}
\usage{
ensemblingModels(ensemble_method = "median", forecasts)
}
\arguments{
\item{ensemble_method}{Character (default = "median"). A ensemble method
("median", "mean").}

\item{forecasts}{A data frame with values to be combined (by rows).}
}
\value{
Combined values.
}
\description{
Combines the k predictions based on the specified method (mean, or median)
}
\examples{
test_df = data.frame(matrix(ncol=3, nrow=5))
colnames(test_df) = c("model_1", "model_2", "model_3")
test_df$model_1 = c(0.25, 0.35, 0.35, 0.23, 0.21)
test_df$model_2 = c(0.27, 0.27, 0.33, 0.35, 0.29)
test_df$model_3 = c(0.29, 0.37, 0.22, 0.37, 0.37)
ensemblingModels(ensemble_method = "median", forecasts = test_df)
}
