% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESTC.R
\name{getBestInstanceModel}
\alias{getBestInstanceModel}
\title{getBestInstanceModel}
\usage{
getBestInstanceModel(df)
}
\arguments{
\item{df}{A data frame whose first column contains "Target" values
and other columns are models predictions.}
}
\value{
a vector with the best model to each instance
}
\description{
Identifies the best model for each of the instances
}
\examples{
test_df = data.frame(matrix(ncol=4, nrow=5))
colnames(test_df) = c("Target", "model_1", "model_2", "model_3")
test_df$Target = c(0.24, 0.28, 0.25, 0.23, 0.28)
test_df$model_1 = c(0.25, 0.35, 0.35, 0.23, 0.21)
test_df$model_2 = c(0.27, 0.27, 0.33, 0.35, 0.29)
test_df$model_3 = c(0.29, 0.37, 0.22, 0.37, 0.37)
getBestInstanceModel(test_df)
}
