% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destc.R
\name{getMSE}
\alias{getMSE}
\title{getMSE}
\usage{
getMSE(target, forecast)
}
\arguments{
\item{target}{Array of targets (observed values).}

\item{forecast}{Array of forecasts (forecasts values).}
}
\value{
Mean squared error
}
\description{
Mean squared error
}
\examples{
target = c(2.7, 2.5, 3.1, 5.6, 5.7)
forecast = c(1.7, 3.5, 2.1, 2.6, 5.1)
getMSE(target,forecast)
}
