% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destc.R
\name{getOrderedListOfBestModelsByTrend}
\alias{getOrderedListOfBestModelsByTrend}
\title{getOrderedListOfBestModelsByTrend}
\usage{
getOrderedListOfBestModelsByTrend(df, tbl_len_perc = 1)
}
\arguments{
\item{df}{A data frame whose first column contains "Target" values,
a 'Model' column which indicates the best model to each instance,
and a 'Class' column which indicate the trend classification
to each instance.}

\item{tbl_len_perc}{float (default = 1). a value that indicates the
percentage of data to be used to count the best models by trend class.}
}
\value{
A list that contains three ranks with the best model to
each class of trend.
}
\description{
Get list with the best models to each trend
}
\examples{
test_df = data.frame(matrix(ncol=4, nrow=10))
colnames(test_df) = c("Target", "model_1", "model_2", "model_3")
test_df$Target = c(0.24, 0.28, 0.25, 0.23, 0.28, 0.24, 0.28, 0.25, 0.23, 0.28)
test_df$model_1 = c(0.25, 0.35, 0.35, 0.23, 0.21, 0.25, 0.35, 0.35, 0.23, 0.21)
test_df$model_2 = c(0.27, 0.27, 0.33, 0.35, 0.29, 0.27, 0.27, 0.33, 0.35, 0.29)
test_df$model_3 = c(0.29, 0.37, 0.22, 0.37, 0.37, 0.29, 0.37, 0.22, 0.37, 0.37)
test_df$Model = best_models; lag = 5
sliding_windows_df = getSlidingWindows_wTrendAnalysis(test_df$Target
                                                   , w = lag, running_mean = 1
                                                   , nStepAhead = 1, alpha = 0.5)
test_df$Class[(lag+1):length(test_df$Target)] = sliding_windows_df$Class
test_df = na.omit(test_df)
getOrderedListOfBestModelsByTrend(test_df, tbl_len_perc = 1)
}
