% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destc.R
\name{getRMSE}
\alias{getRMSE}
\title{getRMSE}
\usage{
getRMSE(target, forecast)
}
\arguments{
\item{target}{Array of targets (observed values).}

\item{forecast}{Array of forecasts (forecasts values).}
}
\value{
RMSE
}
\description{
Root Mean Squared Error (RMSE)
}
\examples{
target = c(2.7, 2.5, 3.1, 5.6, 5.7)
forecast = c(1.7, 3.5, 2.1, 2.6, 5.1)
getRMSE(target,forecast)
}
