% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar.R
\name{getRunningMean}
\alias{getRunningMean}
\title{getRunningMean}
\usage{
getRunningMean(time_series, n)
}
\arguments{
\item{time_series}{a 'ts' or 'numeric' object.}

\item{n}{Int (default = 1). If $running_mean>1$, a running mean function is applied.}
}
\value{
returns a running mean numeric vector.
}
\description{
Calculates the running mean of a time series
}
\examples{
ts = AirPassengers
plot.ts(getRunningMean(ts, 30))
}
