% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar.R
\name{getSlidingWindows_wTrendAnalysis}
\alias{getSlidingWindows_wTrendAnalysis}
\title{getSlidingWindows_wTrendAnalysis}
\usage{
getSlidingWindows_wTrendAnalysis(
  time_series,
  w = 10,
  alpha = 0.1,
  nStepAhead = 1,
  running_mean = 1
)
}
\arguments{
\item{time_series}{a 'ts' or 'numeric' object.}

\item{w}{Int (default = 10). The sliding window lag size.}

\item{alpha}{Float (default = 0.1). The significance level (alpha) of MK trend test.}

\item{nStepAhead}{Int (default = 1). Number o steps ahead .}

\item{running_mean}{Int (default = 1). If running_mean>1, a running mean function is applied.}
}
\value{
A data frame with sliding windows and trend analysis values.
 A and B are Ax+B values from a linear regression, mk_pvalue is the MK trend test
 p-value, S is the MK statists, and Class is the trend classification.
}
\description{
Creates a matrix based on the sliding window technique (w=lag) 
and classifies the trend of each instance based on mk.test().
}
\examples{
ts = AirPassengers
getSlidingWindows_wTrendAnalysis(ts)

}
