% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destc.R
\name{getTheil}
\alias{getTheil}
\title{getTheil}
\usage{
getTheil(target, forecast)
}
\arguments{
\item{target}{Array of targets (observed values).}

\item{forecast}{Array of forecasts (forecasts values).}
}
\value{
Theil's U2 Statistic
}
\description{
Theil's U2 Statistic
}
\examples{
target = c(2.7, 2.5, 3.1, 5.6, 5.7)
forecast = c(1.7, 3.5, 2.1, 2.6, 5.1)
getTheil(target,forecast)
}
